requirements.txt
confluent_kafka
psycopg2-binary
requests
pandas
streamlit
streamlit_autorefresh
plotly
kafka-python